Updated Playground example
Show an example that loads JSON from a file and decodes it into a
struct.
