+++ Argo.xcworkspace/contents.xcworkspacedata
@@ -2,9 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:Argo.playground">
-   </FileRef>
-   <FileRef
       location = "group:Argo.xcodeproj">
    </FileRef>
    <FileRef
@@ -13,4 +10,11 @@
    <FileRef
       location = "group:Carthage/Checkouts/Runes/Runes.xcodeproj">
    </FileRef>
+   <Group
+      location = "group:Playgrounds"
+      name = "Playground examples">
+      <FileRef
+         location = "group:Argo-Basics.playground">
+      </FileRef>
+   </Group>
 </Workspace>
