+++ Playgrounds/Argo-Basics.playground/Resources/user_with_email.json
@@ -1,0 +1,5 @@
+{
+  "id": 1,
+  "name": "Cool User",
+  "email": "u.cool@example.com"
+}
